Refactor out odd stuff
