+++ storj.js
@@ -1,0 +1,3 @@
+export function storj_commands(commands) {
+  // TODO: Care about this!
+}
